requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
Django==4.1.7
django-crispy-forms==2.0
extruct==0.14.0
html-text==0.5.2
html5lib==1.1
idna==3.4
isodate==0.6.1
jstyleson==0.0.2
lxml==4.9.2
mf2py==1.1.2
Pillow==9.4.0
pyparsing==3.0.9
pyRdfa3==3.5.3
rdflib==6.3.2
recipe-scrapers==14.35.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.15
w3lib==2.1.1
webencodings==0.5.1
